+++ bestmobabot/logger.py
@@ -3,12 +3,33 @@ Logger initialisation.
 """
 
 import logging
-from typing import Iterable
+from typing import Iterable, TextIO
+
+import coloredlogs
 
 import bestmobabot.responses
+from bestmobabot.constants import SPAM
 
 
 logger = logging.getLogger('bestmobabot')
+logging.addLevelName(SPAM, 'SPAM')
+
+
+def get_logging_level(verbosity: int) -> int:
+    if verbosity == 0:
+        return logging.INFO
+    if verbosity == 1:
+        return logging.DEBUG
+    return SPAM
+
+
+def install_logging(logger: logging.Logger, verbosity: int, stream: TextIO):
+    coloredlogs.install(
+        get_logging_level(verbosity),
+        fmt='%(asctime)s %(levelname)s %(message)s',
+        logger=logger,
+        stream=stream,
+    )
 
 
 def log_heroes(emoji: str, message: str, heroes: Iterable['bestmobabot.responses.Hero']):